graph_main/requirements.txt
alembic==1.7.7
apache-airflow==1.10.7
apispec==1.3.3
argcomplete==1.12.3
attrs==19.3.0
Babel==2.11.0
cached-property==1.5.2
cattrs==0.9.2
certifi==2022.12.7
chardet==3.0.4
click==7.1.2
colorama==0.4.5
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.37
cycler==0.11.0
Cython==0.29.21
dataclasses==0.8
defusedxml==0.7.1
dill==0.3.4
dnspython==2.2.1
docutils==0.18.1
email-validator==1.3.1
eventlet==0.33.2
funcsigs==1.0.2
future==0.16.0
gevent==21.12.0
graphviz==0.19.1
greenlet==1.1.3.post0
gunicorn==19.10.0
idna==2.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
iso8601==1.1.0
iteration-utilities==0.11.0
itsdangerous==2.0.1
joblib==1.1.1
json-logging==1.2.0
json-merge-patch==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
kombu==4.6.11
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.18.0
mock==4.0.3
natsort==8.2.0
neo4j==4.3.1
nested-lookup==0.2.22
numpy==1.17.1
pandas==0.25.1
path==16.2.0
path.py==12.5.0
pendulum==1.4.4
Pillow==8.4.0
pkg-resources==0.0.0
poppler-utils==0.1.0
prison==0.2.1
psutil==5.9.5
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-daemon==2.1.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
setproctitle==1.2.3
six==1.16.0
SQLAlchemy==1.3.10
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.41.1
tabulate==0.8.10
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.16.0
tqdm==4.64.1
typing==3.7.4.3
typing-extensions==4.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.14.1
WTForms==2.2.1
zipp==3.6.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.5.2
